+++ plugins/moxie/src/types/index.ts
@@ -1,5 +1,7 @@
 import type { BaseProcessResult, CSSPropertyOrVariable } from '@tenoxui/core'
 import type { Plugin } from '@nousantx/plugify'
+import type { StringRules, ArrayRules, ObjectRules } from './utilityRulesResult'
+import type { CreateRegexpResult } from './regexp'
 
 export interface ProcessedValue {
   key: string
@@ -13,7 +15,7 @@ export interface CSSRule {
 }
 
 export interface ProcessResult extends BaseProcessResult<string | ClassNameObject> {
-  use: string
+  use: 'moxie'
   rules: CSSRule | CSSRule[] | string | null
   variant?: string | null
   isImportant?: boolean
@@ -39,10 +41,6 @@ export interface TransformResult {
   }
 }
 
-export interface UtilityFunction {
-  (context: UtilityContext): UtilityResult | null
-}
-
 export interface UtilityContext {
   className: string | ClassNameObject
   value: string | ProcessedValue
@@ -50,18 +48,26 @@ export interface UtilityContext {
   key?: string | null
 }
 
+export type AllowedUtilityRules =
+  | CSSRule
+  | CSSRule[]
+  | StringRules
+  | CSSPropertyOrVariable[]
+  | ArrayRules
+  | ObjectRules
+
 export interface UtilityResult {
   className?: string | ClassNameObject
   property?: CSSPropertyOrVariable | CSSPropertyOrVariable[]
   value?: string
-  rules?:
-    | CSSRule
-    | CSSRule[]
-    | string
-    | [CSSPropertyOrVariable | CSSPropertyOrVariable[], string, boolean?]
-    | { [key: CSSPropertyOrVariable | string]: string | [string, boolean?] }
-
+  rules?: AllowedUtilityRules
   isImportant?: boolean
+}
+
+export type UtilityErrorResult = null | { fail?: true; reason?: string } | [null, string]
+
+export interface UtilityFunction {
+  (context: UtilityContext): UtilityResult | UtilityErrorResult
 }
 
 export interface UtilityConfig {
@@ -83,8 +89,10 @@ export type VariantContext = {
   key?: string | null
   value?: string
 }
+
 export type VariantResult = string | null
-export type VariantFunction = (context?: VariantContext) => VariantResult
+
+export type VariantFunction = (context: VariantContext) => VariantResult
 
 export interface Variants {
   [key: string]: string | VariantFunction
@@ -104,31 +112,7 @@ export interface Config {
   onMatcherCreated?: ((matcher: RegExp) => void) | null
 }
 
-export interface CreatePatternsConfig {
-  variants?: string[]
-  utilities?: string[]
-  safelist?: string[]
-  inputPrefixChars?: string[]
-}
-
-export interface Patterns {
-  variant: string
-  property: string
-  value: string
-}
-
-export interface CreateRegexpResult {
-  patterns: Patterns
-  matcher: RegExp
-  debug: {
-    variantPattern: string
-    propertyPattern: string
-    valuePattern: string
-    fullPattern: string
-  }
-}
-
-type CreateResultContext = (
+export type CreateResultContext = (
   className: string | ClassNameObject,
   variant: string | null,
   property: string | string[],
@@ -138,7 +122,7 @@ type CreateResultContext = (
   fullRules?: any
 ) => ProcessResult
 
-type CreateErrorResultContext = (className: string, reason: string) => InvalidResult
+export type CreateErrorResultContext = (className: string, reason: string) => InvalidResult
 
 export interface ProcessValueContext {
   value: string
@@ -162,10 +146,13 @@ export interface ProcessUtilitiesContext {
 export type ProcessResultFn = ProcessResult | InvalidResult | null
 
 export type ProcessVariantFn = (variant: string, variants?: any) => string | null
+
 export type ProcessValueFn = (ctx: Partial<ProcessValueContext>) => string | null | ProcessedValue
+
 export type ProcessUtilitiesFn = (
   ctx: { className: string | ClassNameObject } & Partial<ProcessUtilitiesContext>
 ) => ProcessResultFn
+
 export type ProcessFn = (ctx: Partial<ProcessContext>) => ProcessResultFn
 
 export interface ProcessContext {
@@ -197,3 +184,6 @@ export interface PluginContext {
     [utilitiesName: string]: Utilities
   }
 }
+
+export * from './regexp'
+export * from './utilityRulesResult'